fix: UI memory leaks, type corrections and build updates

+++ src/AstGUI/Foundation/UiSelectFrame.cpp
@@ -192,7 +192,7 @@ Frame* UiSelectFrame::getSelectedFrame()
     auto* item = frameList_->currentItem();
     if (!item)
         return nullptr;
-    return reinterpret_cast<Frame*>(item->data(Qt::UserRole).value<uintptr_t>());
+    return reinterpret_cast<Frame*>(item->data(Qt::UserRole).value<qulonglong>());
 }
 
 void UiSelectFrame::setSelectedFrame(Frame* frame)
@@ -207,7 +207,7 @@ void UiSelectFrame::setSelectedFrame(Frame* frame)
     for (int i = 0; i < frameList_->count(); ++i)
     {
         auto* item = frameList_->item(i);
-        auto* f = reinterpret_cast<Frame*>(item->data(Qt::UserRole).value<uintptr_t>());
+        auto* f = reinterpret_cast<Frame*>(item->data(Qt::UserRole).value<qulonglong>());
         if (f == frame)
         {
             frameList_->setCurrentRow(i);
@@ -222,7 +222,7 @@ CelestialBody* UiSelectFrame::getSelectedBody() const
     if (!item)
         return nullptr;
     return reinterpret_cast<CelestialBody*>(
-        item->data(Qt::UserRole).value<uintptr_t>());
+        item->data(Qt::UserRole).value<qulonglong>());
 }
 
 void UiSelectFrame::setBody(CelestialBody* body)
@@ -234,7 +234,7 @@ void UiSelectFrame::setBody(CelestialBody* body)
     {
         auto* item = bodyList_->item(i);
         auto* b = reinterpret_cast<CelestialBody*>(
-            item->data(Qt::UserRole).value<uintptr_t>());
+            item->data(Qt::UserRole).value<qulonglong>());
         if (b == body)
         {
             bodyList_->blockSignals(true);
